Implements a tool for building releases and download beta releases
A new tab called "Release" will appear in DEBUG only.
This tab will help updating the manifest.xml and packaging all files
related to the numbered build.

The manifest format has changed entirely and is incompatible with
previous one. A transition from old<>new launcher will be possible
by keeping both separated in the orionuo.com repository.

The new launcher assumes a rigid structure for building releases:

a "release" folder must exist somewhere with all platform specific
manifest xml inside, eg. win64.manifest.xml, osx.manifest.xml, etc.
This folder has the following hierarchy:
  <platform>/<product>/<files...>

where platforms are one of the following:
  win64
  unbuntu
  manjarolinux
  osx

and "-beta" variants.

products are one of the following:
  OrionUO
  OrionLauncher
  OrionAssistant

Both platforms and products are hardcoded, any other value will be
ignored.

When building a release, a product must be chosen and a version string
entered. This version string may be anything and is not checked, it is
only an user-facing information. Entering a existing version string will
overwrite the build.

To reduce bandwidth (server and users) and storage, the builds are
"virtual", composed of specific versions of each file that will exist
only once in the disk (hash addressed).

The input files for a release are hashed and the hash is the real
version information of each individual file. A package is a set of files
at a specific version each. If a new input hash is found, this will
generate an output artifact that is the file itself zipped and stored at
a "update" folder (sibling of the "release" folder).

To the amount of files in the update folder, files are stored in
subdirectories based in the last two bytes of their input hash.

Client side, the launcher will hash all local files and look in the
manifest for a package with version of "incremental". If the hash of the
file mismatch the same on this package, a download will be done.
This file will also be stored in a local cache, so re-downloads will be
reduced (in a per-launcher installation at least).